Introduce base class for pages
This required a bit of reworking, but it makes the code nicer to work with
at least. Currently, things aren't restructured to use the class members,
and the setup wizard will freeze during an installation attempt without
responding to any user input. As such, you will have to kill the process.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>